% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtsne_prep.R
\name{hbeta}
\alias{hbeta}
\title{Estimate perplexity and probability values}
\usage{
hbeta(D, beta = 1)
}
\arguments{
\item{D}{A distance vector.}

\item{beta}{A constant scalar.}
}
\description{
\code{hbeta} returns the perplexity and probability values for a row
   of data \code{D}.
}
