% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtsne_prep.R
\name{p2sp}
\alias{p2sp}
\title{Probability matrix to symmetric probability matrix}
\usage{
p2sp(P)
}
\arguments{
\item{P}{An \code{N x N} probability matrix, like those produced by
\code{\link{x2p}}}
}
\value{
An \code{N x N} symmetrical matrix of pair-wise probabilities.
}
\description{
\code{p2sp} returns a symmetrical pair-wise joint probability
   matrix given an input probability matrix \emph{P}.
}
